refactor(spa): extract KitchenComponent from monolith (T4c) [KAN-118]
Kitchen view (sidebar, recipe grid, recycle bin) and its confirmation
modals (delete recipe, empty bin) extracted to standalone component.
Import/export logic moved to KitchenComponent since it depends on
activeCookbookId. AppComponent drops ~770 lines (467 TS + 713 HTML).
